+++ R/ggwordcloud.R
@@ -19,4 +19,5 @@
 #' @importFrom colorspace
 #'   hex
 #'   HLS
+#' @importFrom stats runif
 NULL