+++ NEWS.md
@@ -1,6 +1,6 @@
 # ggwordcloud 0.1.0.9000
 * add the rm_outside option
-* faster grid.snap
+* better boxes
 * wordcloud wrapper
 
 # ggwordcloud 0.1.0