+++ NEWS.md
@@ -1,7 +1,7 @@
 # ggwordcloud 0.1.0.9000
 * add the rm_outside option
-* better boxes
-* wordcloud and wordcloud2 wrapper
+* wordcloud layout acceleration thanks to better boxing
+* wordcloud and wordcloud2 approximate equivalent
 
 # ggwordcloud 0.1.0
 * Initial version